feat: add Other component for displaying recommended things
This commit adds a new component called Other, which is responsible for displaying recommended things. It takes in a title, an array of data, and a contentKey object to extract the necessary information. The component renders a list of recommended things with their titles and content. This component will be used to display recommended things in various parts of the application.